Make DestinationServiceQuery generic
The goals of this change are:
1. Reduce the size/complexity of `control::discovery` in order to ease
   burden on code reviews.
2. Extract a reusable grpc streaming utility.

There are no intended functional changes.

`control::discovery::DestinationServiceQuery` is used to track the state
of a request (and streaming response) to the destination service. Very
little of this logic is specific to the destination service.

The `DestinationServiceQuery` and associated `UpdateRx` type have been
moved to a new module, `control::remote_stream`, as `Remote` and
`Receiver`, respectively. Both of these types are generic over the gRPC
message type, so it will be possible to use this utility with additional
API endpoints.

The `Receiver::poll` implementation has been simplified to be more
idiomatic with the rest of our code (namely, using `try_ready!`).

Additionally, the enum states are now hidden behind an API. `Remote` now
exposes constructors, a `needs_reconnect()` method that indicates
whether the remote is disconnected, and a `into_receiver_maybe()` that
consumes the remote, producing a receiver if one exists.